tools/Garuda-DOS-Toolkit/requirements.txt
aiohttp
fake-useragent
PyYAML